+++ src/test/java/com/google/api/codegen/gapic/GapicTestModelGenerator.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.api.codegen;
+package com.google.api.codegen.gapic;
 
 import com.google.api.tools.framework.model.Experiments;
 import com.google.api.tools.framework.model.testing.TestConfig;
