+++ src/test/java/com/google/api/codegen/util/ProtoDocumentLinkTest.java
@@ -14,7 +14,6 @@
  */
 package com.google.api.codegen.util;
 
-import com.google.api.codegen.util.CommentPatterns;
 import com.google.api.codegen.util.js.JSCommentReformatter;
 import com.google.api.codegen.util.py.PythonCommentReformatter;
 import com.google.api.codegen.util.ruby.RubyCommentReformatter;
